+++ include/bitcoin/bitcoin/message/inventory_vector.hpp
@@ -54,7 +54,7 @@ public:
         std::istream& stream);
     static inventory_vector factory_from_data(uint32_t version,
         reader& source);
-    static uint64_t satoshi_fixed_size(uint32_t version);
+    static size_t satoshi_fixed_size(uint32_t version);
 
     inventory_vector();
     inventory_vector(type_id type, const hash_digest& hash);
@@ -81,7 +81,7 @@ public:
     void to_data(uint32_t version, writer& sink) const;
     bool is_valid() const;
     void reset();
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     // This class is move assignable but not copy assignable.
     inventory_vector& operator=(inventory_vector&& other);
