+++ src/message/filter_clear.cpp
@@ -140,12 +140,12 @@ void filter_clear::to_data(uint32_t version, writer& sink) const
 {
 }
 
-uint64_t filter_clear::serialized_size(uint32_t version) const
+size_t filter_clear::serialized_size(uint32_t version) const
 {
     return filter_clear::satoshi_fixed_size(version);
 }
 
-uint64_t filter_clear::satoshi_fixed_size(uint32_t version)
+size_t filter_clear::satoshi_fixed_size(uint32_t version)
 {
     return 0;
 }
