+++ src/message/merkle_block.cpp
@@ -173,9 +173,9 @@ void merkle_block::to_data(uint32_t version, writer& sink) const
     sink.write_bytes(flags_);
 }
 
-uint64_t merkle_block::serialized_size(uint32_t version) const
+size_t merkle_block::serialized_size(uint32_t version) const
 {
-    return header_.serialized_size() + 4 +
+    return header_.serialized_size() + 4u +
         variable_uint_size(hashes_.size()) + (hash_size * hashes_.size()) +
         variable_uint_size(flags_.size()) + flags_.size();
 }
